refactor(health): Reorganize UI into feature-specific packages
This commit refactors the Health feature's UI layer by reorganizing composables into a more structured, feature-based package hierarchy. The goal is to improve code organization, modularity, and clarity by grouping files according to their role in the UI (e.g., `overview`, `sessions`, `settings`).

- **New Package Structure**:
    - `components/common`: For generic, app-wide composables like `LoadingScreen` and `ErrorScreen`.
    - `components/charts`: For chart-related composables, such as the extracted `GenericWeeklyChart`.
    - `overview`: Contains the new `OverviewTab` composable, which displays the weekly activity charts.
    - `sessions`: Groups session-related UI, including `SessionsTab`, `SessionCard`, `SessionDetailContent`, and `SessionDetailDialog`.
    - `settings`: For settings-related UI, containing the new `SettingsTab` and the moved `HealthConnectionStatus`.
    - `zout`: A new package to house unused or deprecated composables, cleaning up the main source tree.

- **Refactoring of `HealthDashboard.kt`**:
    - The monolithic `HealthDashboard` has been broken down. The logic and UI for each bottom navigation tab have been extracted into their own dedicated composable files:
        - `OverviewTab.kt`: Manages the display of weekly charts.
        - `SessionsTab.kt`: Manages the list of exercise sessions and the detail dialog logic.
        - `SettingsTab.kt`: Manages the connection status and debug action buttons.
    - `HealthDashboard.kt` now acts as a clean scaffold, orchestrating the display of these tabs.

- **Component Extraction and Relocation**:
    - **`SessionDetailContent.kt` (New)**: The content of the session detail view has been extracted from `SessionDetailScreen.kt` for reusability.
    - **`SessionCard.kt` (New)** and **`DetailRow.kt` (New)**: These new composables were created by extracting UI logic previously embedded within `HealthDashboard.kt`.
    - Composables like `LoadingScreen`, `ErrorScreen`, `HealthConnectionStatus`, `GenericWeeklyChart`, and `SessionDetailDialog` have been moved to their new, more specific packages.

- **Documentation**:
    - `PACKAGE_STRUCTURE.md` has been added to document the new, clean package organization for the health feature.
