+++ features/health/src/main/java/com/zoewave/probase/features/health/ui/components/HealthDataScreen.kt
@@ -28,6 +28,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.zoewave.probase.features.health.ui.HealthEvent
 import com.zoewave.probase.features.health.ui.HealthUiState
+import com.zoewave.probase.features.health.ui.components.charts.GenericWeeklyChart
 
 @Composable
 fun HealthDataScreen(
