--- features/health/src/main/java/com/zoewave/probase/features/health/ui/components/SessionDetailDialog.kt
@@ -1,79 +1,0 @@
-package com.zoewave.probase.features.health.ui.components
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-import androidx.health.connect.client.records.ExerciseSessionRecord
-
-// Update the SessionDetailDialog signature to accept the event handler
-@Composable
-fun SessionDetailDialog(
-    session: ExerciseSessionRecord,
-    onDismiss: () -> Unit,
-    onDelete: (String) -> Unit // <--- NEW CALLBACK
-) {
-    // ... (Formatters and Duration calculation same as before) ...
-
-    Dialog(onDismissRequest = onDismiss) {
-        Card(
-            shape = RoundedCornerShape(16.dp),
-            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface),
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Column(
-                modifier = Modifier.padding(24.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                // ... (Icon, Title, Rows same as before) ...
-
-                Spacer(modifier = Modifier.height(24.dp))
-
-                // --- Action Buttons ---
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    // DELETE BUTTON
-                    OutlinedButton(
-                        onClick = {
-                            onDelete(session.metadata.id)
-                            onDismiss()
-                        },
-                        colors = ButtonDefaults.outlinedButtonColors(
-                            contentColor = MaterialTheme.colorScheme.error
-                        ),
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Text("Delete")
-                    }
-
-                    // CLOSE BUTTON
-                    Button(
-                        onClick = onDismiss,
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Text("Close")
-                    }
-                }
-            }
-        }
-    }
-}
-
