+++ features/health/src/main/java/com/zoewave/probase/features/health/ui/components/common/LoadingScreen.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.ui.components
+package com.zoewave.probase.features.health.ui.components.common
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
