+++ features/health/src/main/java/com/zoewave/probase/features/health/ui/sessions/SessionDetailScreen.kt
@@ -1,11 +1,7 @@
-package com.zoewave.probase.features.health.ui.components
+package com.zoewave.probase.features.health.ui.sessions
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
@@ -31,7 +27,6 @@ import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.zoewave.probase.core.data.service.health.ExerciseSessionData
 import com.zoewave.probase.features.health.ui.HealthViewModel
-import java.time.Duration
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -104,53 +99,7 @@ fun SessionDetailScreen(
     }
 }
 
-@Composable
-private fun SessionDetailContent(
-    data: ExerciseSessionData,
-    modifier: Modifier = Modifier
-) {
-    Column(modifier = modifier.fillMaxWidth(), verticalArrangement = Arrangement.spacedBy(12.dp)) {
-        Text("ID ${data.uid}", style = MaterialTheme.typography.titleLarge)
 
-        Text(
-            "Duration: ${formatDuration(data.totalActiveTime)}",
-            style = MaterialTheme.typography.bodyLarge
-        )
-        data.totalDistance?.let {
-            Text(
-                "Distance: ${"%.2f".format(it.inFeet)} km",
-                style = MaterialTheme.typography.bodyLarge
-            )
-        }
-        data.totalSteps?.let {
-            Text("Steps: $it", style = MaterialTheme.typography.bodyLarge)
-        }
-        data.totalEnergyBurned?.let {
-            Text(
-                "Calories: ${"%.0f".format(it.inCalories)} kcal",
-                style = MaterialTheme.typography.bodyLarge
-            )
-        }
-        Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-            data.minHeartRate?.let { Text("Min HR: $it bpm") }
-            data.avgHeartRate?.let { Text("Avg HR: $it bpm") }
-            data.maxHeartRate?.let { Text("Max HR: $it bpm") }
-        }
-    }
-}
-
-private fun formatDuration(duration: Duration?): String {
-    if (duration == null) return "--"
-    val h = duration.toHours()
-    val m = duration.toMinutes() % 60
-    val s = duration.seconds % 60
-    return buildString {
-        if (h > 0) append("${h}h ")
-        if (m > 0 || h > 0) append("${m}m ")
-        append("${s}s")
-    }
-}
-/*
 /*@Preview(showBackground = true)
 @Composable
 fun SessionDetailScreenPreview() {
