+++ features/health/src/main/java/com/zoewave/probase/features/health/ui/settings/SessionList.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.ui.components
+package com.zoewave.probase.features.health.ui.settings
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.lazy.LazyColumn
@@ -7,6 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.health.connect.client.records.ExerciseSessionRecord
+import com.zoewave.probase.features.health.ui.components.ExerciseSessionRow
 import java.time.ZonedDateTime
