+++ features/health/src/main/java/com/zoewave/probase/features/health/ui/settings/SettingsTab.kt
@@ -1,0 +1,63 @@
+package com.zoewave.probase.features.health.ui.settings
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.zoewave.probase.features.health.ui.HealthEvent
+
+@Composable
+fun SettingsTab(
+    onEvent: (HealthEvent) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = modifier.fillMaxSize()
+    ) {
+        Text(
+            text = "Health Connect Settings",
+            style = MaterialTheme.typography.headlineMedium,
+            modifier = Modifier.padding(bottom = 24.dp)
+        )
+
+        // Re-use your existing connection status component
+        HealthConnectionStatus(onEvent = onEvent)
+
+        Spacer(modifier = Modifier.height(24.dp))
+
+        Text(
+            text = "Debug Tools",
+            style = MaterialTheme.typography.titleMedium,
+            color = MaterialTheme.colorScheme.primary
+        )
+        Spacer(modifier = Modifier.height(8.dp))
+        Text(
+            text = "Use these tools to simulate data if you don't have a real device connected.",
+            style = MaterialTheme.typography.bodyMedium,
+            color = Color.Gray,
+            modifier = Modifier.padding(bottom = 16.dp)
+        )
+
+        Button(
+            onClick = { onEvent(HealthEvent.WriteTestRide) },
+            colors = ButtonDefaults.buttonColors(
+                containerColor = MaterialTheme.colorScheme.tertiary
+            ),
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text("Add Test City Ride (4.5km)")
+        }
+    }
+}
