+++ features/health/src/main/java/com/zoewave/probase/features/health/ui/zout/HealthDataScreen.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.ui.components.notused
+package com.zoewave.probase.features.health.ui.zout
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
