+++ features/health/src/main/java/com/zoewave/probase/features/health/ui/zout/stage/ExerciseSessionInfoColumn.kt
@@ -1,4 +1,4 @@
-package com.zoewave.probase.features.health.ui.components
+package com.zoewave.probase.features.health.ui.zout.stage
 
 ////import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.foundation.clickable
